+++ tests/v1/storage_backend/test_local_cpu_backend.py
@@ -1,7 +1,5 @@
 # SPDX-License-Identifier: Apache-2.0
 # Standard
-from contextlib import nullcontext
-from unittest.mock import patch
 import threading
 
 # Third Party
@@ -13,126 +11,11 @@ from lmcache.utils import CacheEngineKey
 from lmcache.v1.config import LMCacheEngineConfig
 from lmcache.v1.memory_management import (
     AdHocMemoryAllocator,
-    BufferAllocator,
     MemoryFormat,
     MemoryObj,
     MixedMemoryAllocator,
-    PagedTensorMemoryAllocator,
-    TensorMemoryAllocator,
 )
 from lmcache.v1.storage_backend.local_cpu_backend import LocalCPUBackend
-
-
-# This is to mock the constructor and destructor of
-# MixedMemoryAllocator and PinMemoryAllocator to
-# use pin_memory=True for their constructors and
-# avoid calling cudaHostRegister and cudaHostUnregister
-# which may throw an error if torch.empty returns a buffer
-# that cannot be registered (which happens quicker on some machines,
-# especially when torch is doing many allocations and frees)
-@pytest.fixture(autouse=True, scope="module")
-def patch_mixed_allocator():
-    def fake_mixed_init(self, size: int, use_paging: bool = False, **kwargs):
-        """
-        :param int size: The size of the pinned memory in bytes.
-        """
-
-        # self.buffer = torch.empty(size, dtype=torch.uint8)
-        # ptr = self.buffer.data_ptr()
-        # err = torch.cuda.cudart().cudaHostRegister(ptr, size, 0)
-        # assert err == 0, (
-        #     f"cudaHostRegister failed: {torch.cuda.cudart().cudaGetErrorString(err)}"
-        # )
-        self._unregistered = False
-        self.buffer = torch.empty(size, dtype=torch.uint8, pin_memory=True)
-
-        if use_paging:
-            assert "shape" in kwargs, (
-                "shape must be specified for paged memory allocator"
-            )
-            assert "dtype" in kwargs, (
-                "dtype must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.pin_allocator = PagedTensorMemoryAllocator(
-                tensor=self.buffer,
-                shape=kwargs["shape"],
-                dtype=kwargs["dtype"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            self.pin_allocator = TensorMemoryAllocator(self.buffer)
-
-        self.host_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-        self.buffer_allocator = BufferAllocator("cpu")
-
-    def fake_mixed_close(self):
-        if not self._unregistered:
-            torch.cuda.synchronize()
-            # torch.cuda.cudart().cudaHostUnregister(self.buffer.data_ptr())
-            self._unregistered = True
-
-    with (
-        patch(
-            "lmcache.v1.memory_management.MixedMemoryAllocator.__init__",
-            fake_mixed_init,
-        ),
-        patch(
-            "lmcache.v1.memory_management.MixedMemoryAllocator.close", fake_mixed_close
-        ),
-    ):
-        yield
-
-
-@pytest.fixture(autouse=True, scope="module")
-def patch_pin_allocator():
-    def fake_pin_init(self, size: int, use_paging: bool = False, **kwargs):
-        """
-        :param int size: The size of the pinned memory in bytes.
-        """
-
-        # self.buffer = torch.empty(size, dtype=torch.uint8)
-        # ptr = self.buffer.data_ptr()
-        # err = torch.cuda.cudart().cudaHostRegister(ptr, size, 0)
-        # assert err == 0, (
-        #     f"cudaHostRegister failed: {torch.cuda.cudart().cudaGetErrorString(err)}"
-        # )
-        self._unregistered = False
-        self.buffer = torch.empty(size, dtype=torch.uint8, pin_memory=True)
-
-        if use_paging:
-            assert "shape" in kwargs, (
-                "shape must be specified for paged memory allocator"
-            )
-            assert "dtype" in kwargs, (
-                "dtype must be specified for paged memory allocator"
-            )
-            assert "fmt" in kwargs, "fmt must be specified for paged memory allocator"
-            self.allocator = PagedTensorMemoryAllocator(
-                tensor=self.buffer,
-                shape=kwargs["shape"],
-                dtype=kwargs["dtype"],
-                fmt=kwargs["fmt"],
-            )
-        else:
-            self.allocator = TensorMemoryAllocator(self.buffer)
-
-        self.host_mem_lock = threading.Lock() if not use_paging else nullcontext()
-
-    def fake_pin_close(self):
-        if not self._unregistered:
-            torch.cuda.synchronize()
-            # torch.cuda.cudart().cudaHostUnregister(self.buffer.data_ptr())
-            self._unregistered = True
-
-    with (
-        patch(
-            "lmcache.v1.memory_management.PinMemoryAllocator.__init__", fake_pin_init
-        ),
-        patch("lmcache.v1.memory_management.PinMemoryAllocator.close", fake_pin_close),
-    ):
-        yield
 
 
 class MockLookupServer:
